requirements.txt
# python - 3.10.13
circlify==0.15.0
matplotlib==3.10.0
neo4j==5.27.0
networkx==3.2.1
pandas==2.2.3
streamlit_plotly_events==0.0.6
plotly==5.24.1
pyspark==3.5.4
python-dotenv
pyvis==0.3.2
streamlit==1.41.1
#  chart-studio==1.1.0
# streamlit_javascript==0.1.5
# pydot==3.0.4
# graphviz==0.20.3
streamlit_agraph==0.0.45
# pygraphviz
 openai==0.28
